requirements.txt
Flask==2.3.3
python-chess==1.999
stockfish==3.28.0
flask-cors==4.0.0